Optimized Chal Gen Code
