+++ enclave/src/podr2_pri/chal_gen.rs
@@ -83,34 +83,6 @@ pub fn chal_gen(n: i64, chal_id: &Vec<u8>) -> PoDR2Chal {
         };
     }
 
-    let mut rng = thread_rng();
-
-    let range = (n as f64 * 4.6) / 100_f64;
-
-    let mut low = range.floor();
-    let mut high = range.ceil();
-
-    if low < 1_f64 {
-        low = 1_f64;
-    }
-    if high < 1_f64 {
-        high = 1_f64;
-    }
-
-    let between = Range::new(low, high + 1_f64);
-    let n_samples = between.ind_sample(&mut rng) as usize;
-
-    // Choose random blocks
-    let mut n_blocks = sgx_rand::sample(&mut rng, 0..n, n_samples);
-    n_blocks.sort();
-
-    for i in 0..n_samples {
-        let mut rng = thread_rng();
-        let v_between = Range::new(0_i64, i64::MAX);
-        let v = v_between.ind_sample(&mut rng);
-        q_elements.push(QElement { i: n_blocks[i], v });
-    }
-
     let mut chal_ident = CHAL_IDENTIFIER.lock().unwrap();
     let mut time_out: i64;
 
@@ -122,6 +94,7 @@ pub fn chal_gen(n: i64, chal_id: &Vec<u8>) -> PoDR2Chal {
         time_out = schedule_time.timestamp();
 
         chal_ident.chal_id = chal_id.to_vec();
+        q_elements = QElement::get_elements(n);
         chal_ident.q_elements = q_elements.clone();
         chal_ident.time_out = schedule_time.timestamp();
 
@@ -166,7 +139,7 @@ fn post_chal_data() {
         &chal_data,
     );
 
-    // Clear contents of CHAL_DATA 
+    // Clear contents of CHAL_DATA
     chal_data.bloom_filter = BloomFilter::zero();
     chal_data.failed_file_hashes = Vec::new();
     chal_data.chal_id = Vec::new();
