+++ enclave/src/podr2_pri/mod.rs
@@ -1,11 +1,14 @@
 use std::sync::SgxMutex;
 
+use self::chal_gen::{ChalData, ChalIdentifier};
 use alloc::string::{String, ToString};
 use alloc::vec::Vec;
 use serde::{Deserialize, Serialize};
+use sgx_rand::{
+    distributions::{IndependentSample, Range},
+    thread_rng,
+};
 use utils;
-
-use self::chal_gen::{ChalIdentifier, ChalData};
 
 pub mod chal_gen;
 pub mod gen_proof;
@@ -14,10 +17,8 @@ pub mod sig_gen;
 pub mod verify_proof;
 
 lazy_static! (
-    static ref CHAL_IDENTIFIER: SgxMutex<ChalIdentifier> =
-        SgxMutex::new(ChalIdentifier::new());
-    static ref CHAL_DATA: SgxMutex<ChalData> =
-        SgxMutex::new(ChalData::new());
+    static ref CHAL_IDENTIFIER: SgxMutex<ChalIdentifier> = SgxMutex::new(ChalIdentifier::new());
+    static ref CHAL_DATA: SgxMutex<ChalData> = SgxMutex::new(ChalData::new());
 );
 
 #[derive(Serialize, Deserialize)]
@@ -72,6 +73,42 @@ impl EncEncrypt {
 pub struct QElement {
     pub i: i64,
     pub v: i64,
+}
+
+impl QElement {
+    #[inline]
+    pub fn get_elements(n: i64) -> Vec<QElement> {
+        let mut q_elements: Vec<QElement> = vec![];
+
+        let mut rng = thread_rng();
+
+        let range = (n as f64 * 4.6) / 100_f64;
+
+        let mut low = range.floor();
+        let mut high = range.ceil();
+
+        if low < 1_f64 {
+            low = 1_f64;
+        }
+        if high < 1_f64 {
+            high = 1_f64;
+        }
+
+        let between = Range::new(low, high + 1_f64);
+        let n_samples = between.ind_sample(&mut rng) as usize;
+
+        // Choose random blocks
+        let mut n_blocks = sgx_rand::sample(&mut rng, 0..n, n_samples);
+        n_blocks.sort();
+
+        for i in 0..n_samples {
+            let mut rng = thread_rng();
+            let v_between = Range::new(0_i64, i64::MAX);
+            let v = v_between.ind_sample(&mut rng);
+            q_elements.push(QElement { i: n_blocks[i], v });
+        }
+        q_elements
+    }
 }
 
 #[derive(Serialize, Deserialize, Debug)]
